*) Code from Most - Bugfix: Fixed some issues with Lifter RA and encumberance. - Bugfix: Fixed an issue with Flurry RA where it wouldn't do the right damage amount. *) Code from samesame - Added: Support for an NPC only ability to make them immune to CC. "CCImmunity", you can add this to a npctemplate. *) Code from Me - Bugfix: An attempted fix for Trueshot. - Bugfix: When you delve an item with melee speed property you will see the property name and no longer a number. - Bugfix: Fixed an issue with Ichor of the Deep RA where if you fired it, and it didn't meet a requirement i.e. range, it would reset the timer. - Bugfix: Mastery of Concentration now lasts for 30 seconds not 30 milliseconds. - Bugfix: You can now attack players who go linkdead, this fixes an exploit where people used /qtd in RvR. - Bugfix: Removed the chance to increase an item quality from blacksmiths as this is not a live feature, and corrected the messages sent (Thanks to Supgee) - Added: New Language support Czech (Thanks to Palecek) and Spanish (Thanks to Zhulgar) also added new translations for German. (Thanks to Thorne23) - Bugfix: Fixed an issue with Toughness RA not increasing HP.

+++ GameServer/gameobjects/GamePlayer.cs
@@ -8329,7 +8329,7 @@ namespace DOL.GS
 				double enc = (double)Strength;
 				RAPropertyEnhancer ab = GetAbility(typeof(LifterAbility)) as RAPropertyEnhancer;
 				if (ab != null)
-					enc *= 1 + (ab.Amount / 100);
+					enc *= 1 + ((double)ab.Amount / 100);
 				return (int)enc;
 			}
 		}
